Refactor(frontend): Home page
Moved Cultures and Admins to their own files and moved styles to a
style.ts file in order to share the commonly used styles between
components.
