+++ frontend/views/Home/styles.ts
@@ -1,0 +1,24 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  spinner: { top: "50%", position: "relative" },
+  fab: {
+    position: "absolute",
+    margin: 16,
+    right: 0,
+    bottom: 0,
+  },
+
+  view: {
+    flex: 1,
+  },
+
+  inviteModal: {
+    padding: 20,
+    background: "white",
+  },
+
+  list: {
+    flex: 1,
+  },
+});
